Improve ObjectPieceGetter errors, create module
